requirements.txt
Django==4.0.2
djangorestframework==3.13.1
requests==2.31.0
gunicorn==21.2.0
PyJWT==2.8.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
nanoid==2.0.0
pillow==10.3.0
boto3==1.34.79
sendgrid==6.11.0